pyproject.toml
[tool.poetry]
name = "webtronics"
version = "0.1.0"
description = ""
authors = ["user <bav-574@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.99.1"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.17"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
pytest = "^7.4.0"
httpx = "^0.24.1"
mypy = "^1.4.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pydantic = {version = "^1.8", extras = ["email"]}
types-requests = "^2.31.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
